Add pass logic to game of go

+++ games/src/app/go/go.component.ts
@@ -22,22 +22,37 @@ export class GoComponent {
    moveLog: string[] = [];
    board: (Stone | undefined)[][] = Array.from({ length: this.boardDimension }, () => Array(this.boardDimension).fill(undefined));
    columnLetters: string[] = ['A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'J', 'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T']; //19 columns
+   gameEnded: boolean = false;
 
-   constructor(private playerService: PlayerService, private boardService: GoBoardService) {}
+   constructor(private playerService: PlayerService, private boardService: GoBoardService) { }
 
    resetBoard() {
+      this.gameEnded = false;
       this.board = Array.from({ length: this.boardDimension }, () => Array(this.boardDimension).fill(undefined));
    }
 
    public placeStone(row: number, column: number) {
-      let placed: boolean = this.playerService.placeStone(row, column, this.activePlayer, this.board);
-      if (placed) {
-      this.boardService.removeDeadStones(this.blackPlayer, this.whitePlayer, this.board);
-         if (this.activePlayer.color === PieceColor.BLACK) {
-            this.activePlayer = this.whitePlayer;
-         } else {
-            this.activePlayer = this.blackPlayer;
+      if (!this.gameEnded) {
+         let placed: boolean = this.playerService.placeStone(row, column, this.activePlayer, this.board);
+         if (placed) {
+            this.activePlayer.hasPassed = false;
+            this.boardService.removeDeadStones(this.blackPlayer, this.whitePlayer, this.board);
+            this.switchTurn();
          }
+      }
+   }
+
+   public pass() {
+      this.activePlayer.hasPassed = true;
+      this.gameEnded = this.blackPlayer.hasPassed && this.whitePlayer.hasPassed;
+      this.switchTurn()
+   }
+
+   private switchTurn() {
+      if (!this.gameEnded && this.activePlayer.color === PieceColor.BLACK) {
+         this.activePlayer = this.whitePlayer;
+      } else {
+         this.activePlayer = this.blackPlayer;
       }
    }
 }
